Separate out query parser and error type
Started out trying to thread it through a reader but it was suggested
this would be a performance problem. Found I could just parameterize the
difference between parsing a query in a URI context vs a query only
context. The internal function signatures are a bit uglier but they are
still pretty easy to follow.

Still working towards dropping the http-types dependency needed for decoding.