Fix some edge cases, add parser options
Added options to parser to specify the strictness of your parser. So far
the lax option seems to be as lax as some of the major uri parsing
libraries, but only on queries: it takes any character till it gets to

Fixed some edge cases where fragments could not be parsed independently
of eachother. Wasn't using endOfInput properly, so invalid data just got
left unconsumed and a valid input was returned.