Fix bug: RelativeRef serialization dropped port
I was misusing the Maybe monad on the schema such that if the schema was
missing (always the case for RelativeRef), the port would be omitted
because it couldn't get to the code branch that decided whether or not
to drop the port. Instead, because relative refs never have a schema by
definition, you never know if its safe to drop the port and thus should
never drop the port.